Stealth Bar, Coin, and lighting
Adding the basic mechanic of the stealth bar. The coin scripts that I had made were unneccessarily complicated so I simplified those. And then I fixed the lighting of the game a bit so its more logical. Also added a VERY basic mechanic that detects when the guard is too close to the player so we can transition to the game over screen